Spawner / projectile update
Projectile movement and collision / deflect logic consolidated into one script
