--- Assets/_scripts/Enemies/MoveLeft.cs
@@ -1,21 +1,0 @@
-using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-
-public class MoveLeft : MonoBehaviour
-{
-    [SerializeField] private float _speed = 5;
-
-    // Start is called before the first frame update
-    void Start()
-    {
-        
-        
-    }
-
-    // Update is called once per frame
-    void Update()
-    {
-        this.transform.Translate(Vector2.left * _speed * Time.deltaTime);
-    }
-}
