--- Assets/_scripts/Enemies/MoveLeft.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 49e37ea0fdff7fb4f86c1f2b11c4ff2f
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
